requirements.txt
# Core dependencies
strands-agents>=0.1.0
boto3>=1.34.0
numpy>=1.24.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
